Add StateManager.js
StateManager class is used to implement and manage a state design
pattern for the canvas GUI and side menu.

The StateManager contains a instance variable for Grid object.
This is passed to both the canvas and menu states during instantiation
which gives all canvas and menu states access to the same Grid object
in memory.

The StateManager maintains a reference for the current "active" canvas
and menu states and contains a switch method which simultaneously
switches the canvas and menu states.